+++ rbx_dom_weak/CHANGELOG.md
@@ -1,6 +1,7 @@
 # rbx\_dom\_weak Changelog
 
 ## Unreleased
+* `RbxValue` and `RbxValueType` can no longer be matched exhaustively, which enables adding new types without breaking code in the future
 * Added support for `UDim`, `UDim2`, and `Content` value types
 * `PhysicalProperties` is no longer a stub type